docker_javbus/requirements.txt
python-telegram-bot[ext]==20.7
aiohttp==3.9.3